Refactor SpreadsheetSchema to a minimal public interface
SpreadsheetSchema becomes an interface exposing constants, FormatSchema,
Iterable<Column>, and five CSV-style query methods (size, column(int),
column(String), columnIndex, columnName, columnNames). Constructor,
fields, and internal methods (resolve, findColumn, getDataRow,
getOriginColumn, forEachHeaderCell, buildStyles, configureSheet, etc.)
move to SpreadsheetSchemaImpl in schema.internal.

Library callers cast incoming FormatSchema instances to
SpreadsheetSchemaImpl. Internal renames to match interface naming:
columnCount->size, getColumn(int)->column(int), columnIndexByName->columnIndex.
The unused getOriginRow accessor is removed.

Breaking change for 1.8.0: SpreadsheetSchema is now an interface; the
public constructor is removed and internal dispatch methods are moved
to SpreadsheetSchemaImpl.
