+++ src/main/java/io/github/scndry/jackson/dataformat/spreadsheet/schema/internal/BackWriteProjection.java
@@ -13,7 +13,6 @@ import lombok.extern.slf4j.Slf4j;
 
 import io.github.scndry.jackson.dataformat.spreadsheet.schema.Column;
 import io.github.scndry.jackson.dataformat.spreadsheet.schema.ColumnPointer;
-import io.github.scndry.jackson.dataformat.spreadsheet.schema.SpreadsheetSchema;
 
 /**
  * Internal helpers for the SSML back-write scenario — per-cell heap
@@ -43,12 +42,12 @@ public final class BackWriteProjection {
      *  Single source of truth shared with {@code SheetDataBuffer.byteSize()}. */
     public static final int ROW_MEMORY_BYTES = 8;
 
-    /** Cache for {@link #requiresBackWriteScope(SpreadsheetSchema)} keyed on
+    /** Cache for {@link #requiresBackWriteScope(SpreadsheetSchemaImpl)} keyed on
      *  schema identity. WeakHashMap so schemas reclaimed by GC drop their
      *  entry — typical applications hold a single schema for their lifetime,
      *  edge-case ClassLoader reload / dynamically generated schemas do not
      *  leak. */
-    private static final Map<SpreadsheetSchema, Boolean> SCOPE_CACHE =
+    private static final Map<SpreadsheetSchemaImpl, Boolean> SCOPE_CACHE =
             Collections.synchronizedMap(new WeakHashMap<>());
 
     private BackWriteProjection() {
@@ -58,7 +57,7 @@ public final class BackWriteProjection {
      *  ({@link #CELL_MEMORY_BYTES}) plus the row directory entry
      *  ({@link #ROW_MEMORY_BYTES}). */
     public static long innerRowMaxBytes(
-            final SpreadsheetSchema schema, final ColumnPointer arrayPointer) {
+            final SpreadsheetSchemaImpl schema, final ColumnPointer arrayPointer) {
         final List<Column> inners = schema.getColumns(arrayPointer);
         return (long) ROW_MEMORY_BYTES + (long) inners.size() * CELL_MEMORY_BYTES;
     }
@@ -66,7 +65,7 @@ public final class BackWriteProjection {
     /** Projected back-write buffer max for a nested list of
      *  {@code listSize} elements rooted at {@code arrayPointer}. */
     public static long project(
-            final SpreadsheetSchema schema, final ColumnPointer arrayPointer, final int listSize) {
+            final SpreadsheetSchemaImpl schema, final ColumnPointer arrayPointer, final int listSize) {
         if (listSize <= 0) return 0;
         return (long) listSize * innerRowMaxBytes(schema, arrayPointer);
     }
@@ -106,7 +105,7 @@ public final class BackWriteProjection {
     /** Returns true when any record scope (root or nested) has an
      *  outer field of that scope appearing after a child array of the
      *  same scope — the trigger for SSML back-write at that scope. */
-    public static boolean hasOuterFieldAfterList(final SpreadsheetSchema schema) {
+    public static boolean hasOuterFieldAfterList(final SpreadsheetSchemaImpl schema) {
         final Set<ColumnPointer> scopes = new LinkedHashSet<>();
         scopes.add(ColumnPointer.empty());
         for (final Column c : schema) {
@@ -124,7 +123,7 @@ public final class BackWriteProjection {
     }
 
     private static boolean _outerAfterListAt(
-            final SpreadsheetSchema schema, final ColumnPointer scope) {
+            final SpreadsheetSchemaImpl schema, final ColumnPointer scope) {
         boolean seenChildArray = false;
         for (final Column c : schema) {
             if (c == null) continue;
@@ -153,7 +152,7 @@ public final class BackWriteProjection {
      *  one immediate child array — sibling lists share the parent row,
      *  so the second list must back-write into rows already produced by
      *  the first list. */
-    public static boolean hasMultipleSiblingLists(final SpreadsheetSchema schema) {
+    public static boolean hasMultipleSiblingLists(final SpreadsheetSchemaImpl schema) {
         final Map<ColumnPointer, Set<ColumnPointer>> childArraysByParent = new HashMap<>();
         for (final Column c : schema) {
             if (c == null) continue;
@@ -173,11 +172,11 @@ public final class BackWriteProjection {
         return false;
     }
 
-    /** Cached form of {@link #hasOuterFieldAfterList(SpreadsheetSchema)} or
-     *  {@link #hasMultipleSiblingLists(SpreadsheetSchema)}. Writers call this
+    /** Cached form of {@link #hasOuterFieldAfterList(SpreadsheetSchemaImpl)} or
+     *  {@link #hasMultipleSiblingLists(SpreadsheetSchemaImpl)}. Writers call this
      *  once per nested array to gate flush suspension — the cache avoids
      *  recomputing the result across every record. */
-    public static boolean requiresBackWriteScope(final SpreadsheetSchema schema) {
+    public static boolean requiresBackWriteScope(final SpreadsheetSchemaImpl schema) {
         final Boolean cached = SCOPE_CACHE.get(schema);
         if (cached != null) return cached;
         final boolean v = hasOuterFieldAfterList(schema) || hasMultipleSiblingLists(schema);
@@ -187,7 +186,7 @@ public final class BackWriteProjection {
 
     /** Logs a warn at schema build time when {@link #hasOuterFieldAfterList}
      *  returns true. Call once after schema construction. */
-    public static void warnIfScenario(final SpreadsheetSchema schema) {
+    public static void warnIfScenario(final SpreadsheetSchemaImpl schema) {
         if (!hasOuterFieldAfterList(schema)) return;
         log.warn(
                 "Schema has an outer field declared after a List<T> field —"
